connected import report with actual errors
