+++ src/main/java/seedu/address/logic/commands/ImportCommand.java
@@ -3,6 +3,7 @@ package seedu.address.logic.commands;
 import static java.util.Objects.requireNonNull;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 import seedu.address.commons.core.EventsCenter;
@@ -10,6 +11,7 @@ import seedu.address.commons.events.ui.ShowHelpRequestEvent;
 import seedu.address.commons.events.ui.ShowImportReportEvent;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.logic.converters.ImportError;
 import seedu.address.logic.converters.PersonConverter;
 import seedu.address.logic.converters.exceptions.PersonDecodingException;
 import seedu.address.logic.converters.fileformats.AdaptedPerson;
@@ -34,6 +36,7 @@ public class ImportCommand extends Command {
 
     private SupportedFile supportedFile;
     private final PersonConverter personConverter;
+    private List<ImportError> errors;
     private int successfulImports;
     private int totalImports;
 
@@ -45,6 +48,7 @@ public class ImportCommand extends Command {
                 : "supportedFile and personConverter does not support the same file format";
         this.supportedFile = supportedFile;
         this.personConverter = personConverter;
+        errors = new ArrayList<>();
     }
 
     @Override
@@ -59,7 +63,10 @@ public class ImportCommand extends Command {
         } catch (IOException ioe) {
             throw new CommandException(ioe.getMessage(), ioe);
         }
-        EventsCenter.getInstance().post(new ShowImportReportEvent());
+
+        if (!errors.isEmpty()){
+            EventsCenter.getInstance().post(new ShowImportReportEvent(errors));
+        }
         model.commitAddressBook();
         return new CommandResult(
                 String.format(MESSAGE_IMPORT_CSV_RESULT, successfulImports, totalImports, supportedFile.getFileName()));
@@ -74,8 +81,10 @@ public class ImportCommand extends Command {
                 Person toAdd = personConverter.decodePerson(person);
                 addPerson(toAdd, model);
             } catch (PersonDecodingException pe) {
+                errors.add(new ImportError(person.getFormattedString(), pe.getMessage()));
                 successfulImports--;
             } catch (CommandException ce) {
+                errors.add(new ImportError(person.getFormattedString(), ce.getMessage()));
                 successfulImports--;
             }
         }
